Dev: admin dashboard features

+++ components/Product/ProductActions.tsx
@@ -84,6 +84,9 @@ export function ProductActions({ productId, product }: ProductActionsProps) {
     if (!session) {
       setIsModalOpen(true);
     } else {
+      if (session?.user.role)
+        return toast.error('Admin cannot add products to wishlist');
+
       setIsModalOpen(false);
       if (userWishlistProductIdArr.includes(productId) && product) {
         users.deleteProductFromWishlist(
